+++ gpioone/__init__.py
@@ -9,3 +9,4 @@ from .dc_motor import *
 from .i2c import *
 from .pca9685 import *
 from .ssd1306 import *
+from .pcd8544 import *